# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Campus Second-hand Market API
  version: 0.1.0-draft
  description: |
    Field-level OpenAPI draft for the campus second-hand market backend.
    This document aligns with team_docs/概要设计v1.4.html section 5 and the current FastAPI router stubs.
    It is intended to be the single source of truth for frontend-backend parallel development.
  contact:
    name: Campus Second-hand Market Team
servers:
  - url: https://api.example.com
    description: Production placeholder
  - url: http://localhost:8000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/orders:
    post:
      tags: [Order]
      summary: Create order request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: Order created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
  /api/orders/{orderId}:
    get:
      tags: [Order]
      summary: Get order detail
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Order'
  /api/orders/{orderId}/seller-confirm:
    post:
      tags: [Order]
      summary: Seller confirms order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/SimpleOk'
  /api/orders/{orderId}/cancel:
    post:
      tags: [Order]
      summary: Cancel order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/SimpleOk'
  /api/orders/{orderId}/complete:
    post:
      tags: [Order]
      summary: Complete order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      responses:
        '200':
          $ref: '#/components/responses/SimpleOk'
  /api/orders/{orderId}/reviews:
    post:
      tags: [Order]
      summary: Submit review for order
      parameters:
        - $ref: '#/components/parameters/OrderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [score, content]
              properties:
                score:
                  type: integer
                  minimum: 1
                  maximum: 5
                content:
                  type: string
      responses:
        '200':
          description: Review submission result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        properties:
                          orderId:
                            type: integer
                          score:
                            type: integer
                          content:
                            type: string
  /api/reports:
    post:
      tags: [Report]
      summary: Create report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
      responses:
        '200':
          description: Report created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Report'
  /api/reports/{reportId}:
    get:
      tags: [Report]
      summary: Get report processing status
      parameters:
        - $ref: '#/components/parameters/ReportId'
      responses:
        '200':
          description: Report detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Report'
  /api/appeals:
    post:
      tags: [Report]
      summary: Submit appeal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetType, targetId, reason]
              properties:
                targetType:
                  type: string
                  enum: [report, user_status]
                targetId:
                  type: integer
                reason:
                  type: string
      responses:
        '200':
          description: Appeal submission result.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    required: [data]
                    properties:
                      data:
                        type: object
                        properties:
                          submitted:
                            type: boolean
                          targetType:
                            type: string
                          targetId:
                            type: integer
                          reason:
                            type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderId:
      name: orderId
      in: path
      required: true
      description: Order identifier.
      schema:
        type: integer
      example: 5001
    ReportId:
      name: reportId
      in: path
      required: true
      description: Report identifier.
      schema:
        type: integer
      example: 7001
  responses:
    SimpleOk:
      description: Generic success response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          examples:
            success:
              value:
                code: 0
                message: ok
                data: {}
                requestId: req-demo-1
                timestamp: '2026-05-14T08:00:00Z'
...
  schemas:
    ApiResponse:
      type: object
      description: Standard API envelope used by every endpoint.
      required: [code, message, requestId, timestamp]
      properties:
        code:
          type: integer
          description: Business code, 0 means success.
          example: 0
        message:
          type: string
          description: Human-readable message.
          example: ok
        data:
          description: Response payload, nullable for empty operations.
          nullable: true
        requestId:
          type: string
          description: Unique request trace identifier.
          example: req-demo-1
        timestamp:
          type: string
          format: date-time
          description: UTC timestamp generated by the backend.
          example: '2026-05-14T08:00:00Z'
...